network-automation-gitops/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
pydantic==2.0.0
python-multipart==0.0.6
pyyaml==6.0.1
gitpython==3.1.40
paramiko==3.3.1
netmiko==4.3.0
napalm==4.1.0
nornir==3.4.0
nornir-netmiko==1.0.0
nornir-napalm==0.2.0
celery==5.3.0
redis==5.0.0
sqlalchemy==2.0.0
alembic==1.11.0
psycopg2-binary==2.9.0
structlog==23.1.0
python-json-logger==2.0.7
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.0
websockets==11.0.3
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0
questionary==2.0.1
inquirer==3.1.3
halo==0.0.31
